create IntDegree.kt and LinearDegree.kt
